Modernize JAX sharding helpers
