+++ dataloader/cudaloader.py
@@ -1,33 +1,21 @@
-#作者：Yuwei
-#链接：https://www.zhihu.com/question/307282137/answer/1560137140
-#来源：知乎
-from threading import Thread
 from queue import Queue
-import jax
-import jax.numpy as jnp
-import numpy as np
-from src.jax_sharding import device_put_sharded
+from threading import Thread
+
+from src.jax_sharding import device_put_leading_axis_sharded, get_jax_devices, leading_axis_sharding
+
 
 class CudaDataLoader:
-
     def __init__(self, loader, queue_size=2):
         self.idx = 0
         self.loader = loader
         self.queue = Queue(maxsize=queue_size)
-        self.val_train=0
-        self.devices = jax.local_devices()
-        self.num_devices = len(self.devices)
-        if self.num_devices != self.loader.local_size:
-            raise RuntimeError(
-                "CudaDataLoader device count does not match loader.local_size: "
-                f"{self.num_devices} vs {self.loader.local_size}. Check "
-                "config.local_size and CUDA_VISIBLE_DEVICES."
-            )
+        self.val_train = 0
+        self.devices = get_jax_devices(loader.local_size)
+        self.sharding = leading_axis_sharding(self.devices)
         self.worker = Thread(target=self.load_loop, daemon=True)
         self.worker.start()
 
     def load_loop(self):
-        # The loop that will load into the queue in the background
         while True:
             for sample in self.loader:
                 sample_set = (sample[0], self._transfer_to_gpus(sample[1:]))
@@ -41,43 +29,18 @@ class CudaDataLoader:
             self.queue.join()
             self.worker.join()
             raise StopIteration
-        elif self.idx > self.loader.ntrain-0.5 and self.val_train < 0.5:
-            self.val_train=1
-            #print(self.val_train, "hello", flush=True)
+        elif self.idx > self.loader.ntrain - 0.5 and self.val_train < 0.5:
+            self.val_train = 1
             raise StopIteration
-        elif self.idx > self.loader.numpoint-0.5 and self.val_train > 0.5:
-            self.val_train=0
+        elif self.idx > self.loader.numpoint - 0.5 and self.val_train > 0.5:
+            self.val_train = 0
             self.idx = 0
             raise StopIteration
         else:
             out = self.queue.get()
             self.idx = out[0]
-            #print(self.idx, "get", flush=True)
             self.queue.task_done()
         return out[1]
 
     def _transfer_to_gpus(self, data):
-        """
-        使用JAX的惯用方法，将一个已经包含设备轴的Pytree分发到各设备。
-
-        Args:
-            data: 一个Pytree，其叶子节点的形状为 (num_devices, ...)。
-                  例如: (a[2, 10, ...], (b[2, 10], c[2, 10]))
-
-        Returns:
-            一个单一的 ShardedDeviceArray 对象，其内部数据分布在所有设备上。
-        """
-        # 步骤1: 将一个 "Pytree of Arrays" 转置成一个 "List of Pytrees"。
-        # 每个Pytree代表一个设备的数据。
-        # jax.tree.map 会帮我们处理所有嵌套结构。
-        list_of_shards = [
-            jax.tree.map(lambda leaf: leaf[i], data)
-            for i in range(self.num_devices)
-        ]
-        # 例如，list_of_shards[0] 会是：
-        # (a[0], (b[0], c[0]))
-        # 其中 a[0] 的形状是 (10, ...), b[0] 的形状是 (10,)
-
-        # 步骤2: 使用 jax.device_put_sharded 创建一个 ShardedDeviceArray。
-        # 这个函数是关键，它会保留数据的全局形状信息。
-        return device_put_sharded(list_of_shards, self.devices)
+        return device_put_leading_axis_sharded(data, self.sharding)
